Improved error handling and checks if database exists now

+++ tvi_dbutils.py
@@ -16,7 +16,7 @@ def get_ips_from_db(conn: mariadb.Connection) -> list[str] | None:
         SELECT number, ip_address FROM number_ip_mappings;
         """
 
-        result = cursor.execute(query)
+        cursor.execute(query)
         records = cursor.fetchall()
         return records
     except mariadb.Error as e:
@@ -34,14 +34,11 @@ def add_numbers_to_db(conn: mariadb.Connection,
 
             if combo.is_valid() == False:
                 logger.error(
-                    "Skipping, Invalid IP/Number format: '%s'" %
-                    combo.get_combo_str(),
-                    exc_info=combo.get_error())
+                    "Skipping, Invalid IP/Number format: '%s'" % combo.get_combo_str(), exc_info=combo.get_error())
                 continue
 
-            data_list.append(
-                (combo.get_raw_ip_address(),
-                 combo.get_phone_number()))
+            data_list.append((combo.get_raw_ip_address(),combo.get_phone_number()))
+
 
         if len(data_list) < 1:
             logger.warning("No numbers were added as all failed to validate")
@@ -60,19 +57,30 @@ def add_numbers_to_db(conn: mariadb.Connection,
         logger.error("MariaDB error: ", exc_info=e)
 
 
-def remove_numbers_from_db(conn: mariadb.Connection, numbers_list: list[int]):
+def remove_numbers_from_db(conn: mariadb.Connection, numbers_list: list[str], number_length=4):
     try:
         cursor = conn.cursor()
 
         data_list = []
 
         for number in numbers_list:
-            data_list.append(tuple([number]))
+            if len(number) != number_length:
+                logger.error("Skipping, number '%s' of length '%s' gotten, expected number of length '%s'" % (number, len(number), number_length))
+                continue
+            
+            data_list.append((number,))
+
+        if len(data_list) < 1:
+            logger.warning("No numbers were removed as none where the correct format")
+            return
+
 
         query = """
         DELETE FROM number_ip_mappings
         WHERE number = ?;
         """
+
+        print(data_list)
 
         cursor.executemany(query, data_list)
 
@@ -105,16 +113,14 @@ def resolve_number_to_ip(conn: mariadb.Connection,
         cursor = conn.cursor()
 
         query = """
-        SELECT ip FROM number_ip_mappings WHERE number = ?;
+        SELECT ip_address FROM number_ip_mappings WHERE number = ?;
         """
         result = cursor.execute(query, (number,))
         result = cursor.fetchone()
 
         if isinstance(result, tuple):
             ip = str(ipaddress.ip_address(result[0]))
-            logger.info(
-                "Resolved number '%s' to IP adress '%s'" %
-                (number, ip))
+            logger.info("Resolved number '%s' to IP adress '%s'" % (number, ip))
             return ip
         return result
     except Exception as e:
@@ -137,7 +143,7 @@ def create_db(user: str, password: str, number_length: int) -> None:
 
         query = """
             CREATE TABLE IF NOT EXISTS number_ip_mappings (
-                number INT UNSIGNED PRIMARY KEY,
+                number VARCHAR(10) PRIMARY KEY,
                 ip_address BINARY(4) NOT NULL,
                 port INT UNSIGNED NULL,
                 UNIQUE (number),
@@ -162,8 +168,26 @@ def create_db(user: str, password: str, number_length: int) -> None:
         cursor.execute(query3, tuple([number_length]))
 
         conn.commit()
+        conn.close()
     except mariadb.Error as e:
         logger.error("MariaDB error: ", exc_info=e)
+
+def database_exists(user, password) -> bool:
+    try:
+        conn = mariadb.connect(host="localhost", user=user, password=password)
+        cursor = conn.cursor()
+        cursor.execute(f"SELECT SCHEMA_NAME FROM INFORMATION_SCHEMA.SCHEMATA WHERE SCHEMA_NAME = 'tvi'")
+        result = cursor.fetchone()
+        conn.close()
+
+        if result:
+            return True
+        else:
+            return False
+
+    except mariadb.Error as e:
+        logger.error("MariaDB Error: ", exc_info=e)
+        return False
 
 
 def drop_db(conn: mariadb.Connection) -> None:
